rename of tree_config.py_trees
