+++ beams/tree_config/__init__.py
@@ -42,14 +42,14 @@ from .base import (BaseItem, BehaviorTreeItem, ExternalItem, PVTarget, Target,
 from .condition import (BaseConditionItem, BinaryConditionItem,
                         ConditionOperator, DummyConditionItem,
                         RangeConditionThing)
-from .pytrees import (BlackboardToStatusItem,
-                      CheckBlackboardVariableExistsItem,
-                      CheckBlackboardVariableValueItem, DummyItem, FailureItem,
-                      PeriodicItem, RunningItem, SetBlackboardVariableItem,
-                      StatusQueueItem, SuccessEveryNItem, SuccessItem,
-                      TickCounterItem, UnsetBlackboardVariableItem,
-                      WaitForBlackboardVariableItem,
-                      WaitForBlackboardVariableValueItem)
+from .py_trees import (BlackboardToStatusItem,
+                       CheckBlackboardVariableExistsItem,
+                       CheckBlackboardVariableValueItem, DummyItem, FailureItem,
+                       PeriodicItem, RunningItem, SetBlackboardVariableItem,
+                       StatusQueueItem, SuccessEveryNItem, SuccessItem,
+                       TickCounterItem, UnsetBlackboardVariableItem,
+                       WaitForBlackboardVariableItem,
+                       WaitForBlackboardVariableValueItem)
 from .tree_config import (BaseSequenceItem, CheckAndDoItem, IncPVActionItem,
                           ParallelItem, ParallelMode, SelectorItem,
                           SequenceConditionItem, SequenceItem, SetPVActionItem,
